Update examples and README for better compatibility and type safety
- Include Arduino.h explicitly in all examples and README snippets
- Change pin constants from int to uint8_t to be more memory efficient
- Standardize indentation in the README code blocks

+++ examples/04-init_and_stop/04-init_and_stop.ino
@@ -1,7 +1,8 @@
+#include <Arduino.h>
 #include <dv_led_blink.h>
 
-const int ledPin = 13;
-const int buttonPin = 2;
+const uint8_t ledPin = 13;
+const uint8_t buttonPin = 2;
 DV_LedBlink led;
 
 void setup() {
